Get initial display of data with the Root item
