+++ src/pages/script/index.tsx
@@ -1,5 +1,13 @@
-import { ReactNode, useEffect, useRef, useState } from 'react';
-import ScriptPage from '@site/src/components/gameScriptComponents/ScriptPage';
+import { ReactNode, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import {
+  ScriptPageEventHandlers,
+  ScriptPageState,
+  ScriptData,
+  ScriptLayout,
+  TableOfContents,
+  RoleTable,
+} from '@site/src/components/gameScriptComponents/ScriptPage';
+import PageRoot from '@site/src/components/pages/Root';
 import Ajv from 'ajv';
 import scriptSchema from './script.schema.json';
 import { GameScript } from '@site/src/components/gameScriptComponents/scriptTypes';
@@ -9,6 +17,8 @@ import LayoutProvider from '@theme/Layout/Provider';
 import Navbar from '@theme/Navbar';
 import pageStyles from './page.module.css';
 import useIsBrowser from '@docusaurus/useIsBrowser';
+import { produce } from 'immer';
+import { createIndex } from '@site/src/components/gameScriptComponents/scriptIndex';
 
 const ajv = new Ajv({
   formats: {
@@ -63,10 +73,102 @@ function useElementSize(
   return [size, setSizedElem];
 }
 
+
+function ScriptPage({ script, fragment }: {
+  script: GameScript,
+  fragment?: string,
+}): ReactNode {
+  useEffect(() => {
+    if (!fragment) {
+      console.log("No fragment")
+      return;
+    }
+
+    const element = document.getElementById(fragment);
+    if (!element) {
+      return;
+    }
+    element?.scrollIntoView();
+  }, [fragment])
+  const [scriptState, setScriptState] = useState<ScriptPageState | null>({});
+  const onFocusClose = useCallback((field: 'role_id' | 'room_id') => {
+    switch (field) {
+      case 'role_id':
+        setScriptState(produce(draft => {
+          draft.focuses = draft.focuses || {};
+          draft.focuses.role_id = null;
+        }));
+        break;
+      case 'room_id':
+        setScriptState(produce(draft => {
+          draft.focuses = draft.focuses || {};
+          draft.focuses.room_id = null;
+        }));
+        break;
+    }
+  }, []);
+  const onRoleSelect = useCallback((role_id: string) => {
+    setScriptState(produce(draft => {
+      draft.focuses = draft.focuses || {};
+      draft.focuses.role_id = role_id;
+    }));
+  }, []);
+  const onRoomSelect = useCallback((room_id: string) => {
+    setScriptState(produce(draft => {
+      draft.focuses = draft.focuses || {};
+      draft.focuses.room_id = room_id;
+    }));
+  }, []);
+
+  const scriptIndex = useMemo(() => createIndex(script), [script]);
+
+  const conversations = useMemo(() => {
+    if (!script) {
+      return [];
+    }
+
+    let conversations = Object.values(scriptIndex.conversations);
+    if (scriptState.focuses?.room_id) {
+      conversations = conversations.filter(conv => {
+        return conv.parentRoom.id === scriptState.focuses.room_id;
+      });
+    }
+
+    if (scriptState.focuses?.role_id) {
+      conversations = conversations.filter(conv => {
+        return conv.containsRole(scriptState.focuses.role_id);
+      });
+    }
+    return conversations;
+  }, [script, scriptState]);
+
+  if (!script) {
+    return null;
+  }
+
+  const sidebar = (
+    <TableOfContents
+      focuses={scriptState.focuses || {}}
+      onFocusClose={onFocusClose}
+      onRoleSelect={onRoleSelect}
+      onRoomSelect={onRoomSelect}
+    />
+  );
+
+  return <ScriptData.Provider value={scriptIndex}>
+    <PageRoot sidebar>
+      <div>
+        <h2>Roles</h2>
+        <RoleTable />
+      </div>
+      <ScriptLayout convs={conversations} />
+    </PageRoot>
+  </ScriptData.Provider>;
+}
+
 export default function Page({ }): ReactNode {
   // We represent the current state as a fragment of the URL.
   const location = useLocation();
-  const [headerSize, setHeaderRef] = useElementSize();
 
   let hash = null;
   if (location.hash) {
@@ -79,17 +181,9 @@ export default function Page({ }): ReactNode {
     console.error(validate.errors);
     throw new Error('Invalid script data');
   }
-  return (<LayoutProvider>
-    <div className={pageStyles.screen}>
-      <div className={pageStyles.navBar} ref={setHeaderRef}>
-        <Navbar />
-      </div>
-      <div className={pageStyles.main}>
-        <ScriptPage
-          script={scriptData as GameScript}
-          headerHeight={headerSize ? headerSize.height : 0}
-          fragment={hash} />
-      </div>
-    </div>
-  </LayoutProvider>)
+  return (
+    <ScriptPage
+      script={scriptData as GameScript}
+      fragment={hash} />
+  );
 }
